+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ### Changelog
 
-### 0.0.2 2 - Aug 2019
+### 0.0.4 - 3 Sep 2019
+- Add `probe` and 'isIt' static methods in `Constants` to find properties of movie or showtimes.
+- Add `getMessage` static method in `Constants`
+
+### 0.0.3.8 - 2 Aug 2019
 - Updated two more Info3 bits  
 
 ### 0.0.1